% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpdemo-package.R
\docType{package}
\name{rpdemo-package}
\alias{rpdemo}
\alias{rpdemo-package}
\title{rpdemo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stephen Turner \email{first.last@example.com}

}
\keyword{internal}
