% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppress_counts}
\alias{suppress_counts}
\title{Suppress low counts}
\usage{
suppress_counts(x, threshold = 5)
}
\arguments{
\item{x}{a numeric vector of counts}

\item{threshold}{numeric threshold for the number below which to suppress}
}
\value{
A vector with low counts suppressed.
}
\description{
This suppresses values with low counts for privacy reasons etc.
}
\examples{
suppress_counts(1:10)
suppress_counts(1:20, threshold=10)
}
